Add Some Enhancement To Datinf App

+++ Api/src/Core/DatingApp.Application/Interfaces/ILikeAppService.cs
@@ -5,7 +5,7 @@ namespace DatingApp.Application.Interfaces
 {
     public interface ILikeAppService
     {
-        Task AddLikeAsync(long sourceUserId, string likedUserName);
-        Task<List<LikeDto>> GetUserLikeAsync(string predicate, long userId);
+        Task AddLikeAsync(int sourceUserId, string likedUserName);
+        Task<List<LikeDto>> GetUserLikeAsync(string predicate, int userId);
     }
 }
