+++ Api/src/Core/DatingApp.Domain/Aggregates/AppUser/Entities/UserLike.cs
@@ -3,8 +3,8 @@
 public class UserLike
 {
     public AppUser SourceUser { get; set; }
-    public long SourceUserId { get; set; }
+    public int SourceUserId { get; set; }
     public AppUser TargetUser { get; set; }
-    public long TargetUserId { get; set; }
+    public int TargetUserId { get; set; }
 
 }
