+++ Api/src/Core/DatingApp.Domain/Aggregates/AppUser/Entities/UserPhoto.cs
@@ -7,7 +7,7 @@ namespace DatingApp.Domain.Aggregates.AppUser.Entities
         public string Url { get; set; } = string.Empty;
         public bool IsMain { get; set; }
         public string? PublicId { get; set; }
-        public long AppUserId { get; set; }
+        public int AppUserId { get; set; }
         public AppUser AppUser { get; set; }
     }
 }
